backend-fastapi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
redis==5.0.4
httpx==0.27.0
pydantic==2.7.0
pydantic[email]==2.7.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
